refactor: extract GitHubIssueExportTarget.swift from IssueExportTargets.swift
Splits the `GitHubIssueExportTarget` Codable target out of
IssueExportTargets.swift.

Byte-identical move of lines 3-30. Both files import Foundation.

IssueExportTargets.swift: 124 → 95 lines. Tests: 667.
